refactor: combine `src/` and `include/`
