+++ src/iguana/services/Object.cc
@@ -1,4 +1,4 @@
-#include "iguana/Object.h"
+#include "Object.h"
 
 namespace iguana {
