+++ src/iguana/services/Object.h
@@ -3,7 +3,7 @@
 #include <string>
 #include <memory>
 
-#include "iguana/Logger.h"
+#include "iguana/services/Logger.h"
 
 namespace iguana {
