+++ src/iguana/services/meson.build
@@ -1,7 +1,11 @@
 services_sources = [
   'Logger.cc',
   'Object.cc',
-  'Algorithm.cc',
+]
+
+services_public_headers = [
+  'Logger.h',
+  'Object.h',
 ]
 
 services_lib = shared_library(
@@ -13,3 +17,9 @@ services_lib = shared_library(
   install_rpath:       project_lib_rpath,
 )
 project_libs += services_lib
+
+install_headers(
+  services_public_headers,
+  subdir:        meson.project_name() / 'services',
+  preserve_path: true,
+)
